requirements.txt
numpy==1.15.2
Pillow==5.3.0
pycrypto==2.6.1
PyQt5==5.9.1
PyQt5-sip==4.19.12
PyWavelets==1.0.1
qrcode==6.0
sip==4.19.8
six==1.11.0
